fix: avoid dynamic imports
